requirements.txt
Appium-Python-Client==2.6.2
async-generator==1.10
attrs==22.1.0
certifi==2022.9.14
et-xmlfile==1.1.0
h11==0.13.0
idna==3.4
iniconfig==1.1.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
